Add autodetection code for VRAM and palettes
VRAM will be autodetected by checking if we're in the Giygas battle which uses the VRAM at 0x3000, or if we're in Mode 0 (only true when 2x 2BPP backgrounds, which is when the VRAM at 0x3000 is used).
Safe palettes will be autodetected by getting the current enemy group, finding the number of enemy types in the group, and using the rest of the palettes not used by the enemies.